+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "dev": "nodemon src/socket-back.js"
+    "dev": "nodemon server/socket-back.js"
   },
   "author": "Denisson Freitas",
   "license": "ISC",
