+++ public/documento/documento.html
@@ -21,7 +21,7 @@
     <textarea id="editor-texto" class="form-control mb-4" style="height: 400px; resize: none;"></textarea>
 
     <div class="d-flex justify-content-between">
-      <a href="index.html" class="btn btn-primary">Voltar</a>
+      <a href="../index.html" class="btn btn-primary">Voltar</a>
       <button id="excluir-documento" class="btn btn-outline-danger">Excluir documento</button>
     </div>
   </main>
