+++ public/documento/socket-documento.js
